# TANDEM Edge/TANDEM Edge Platform Manager (Pi-edge)/edge-implementation/pi-edge-implementation/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: pi-edge Controller API
  description: |
    API exposed by pi-edge for "PaaS & Service Function" - based interaction with NFV MANO.
  termsOfService: http://swagger.io/terms/
  contact:
    email: nikpsarom@intracom-telecom.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
...
servers:
- url: https://virtserver.swaggerhub.com/piedge-connector/2.0.0
paths:
...
  /deployServiceFunction:
    post:
      tags:
      - Service Functions Instances
      summary: Request to deploy a Service function (from the catalogue) to an edge
        node.
      operationId: deploy_service_function
      requestBody:
        description: Deploy Service Function.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/deployServiceFunction'
      responses:
        "200":
          description: App deployed.
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
...
  /chain:
    post:
      tags:
      - Service Functions Instances
      summary: Request to deploy a chain of function services.
      operationId: deploy_chain
      requestBody:
        description: Deploy chain.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/deployChain'
      responses:
        "200":
          description: App deployed.
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
    patch:
      tags:
      - Service Functions Instances
      summary: Request to update a chain of function services.
      operationId: update_chain
      requestBody:
        description: Deploy chain.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/deployChain'
      responses:
        "200":
          description: App deployed.
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
  /updateDeployedServiceFunction:
    patch:
      tags:
      - Service Functions Instances
      summary: Request to update the status of a service.
      operationId: update_deployed_service_function
      requestBody:
        description: update a running service function.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/deployServiceFunction'
      responses:
        "200":
          description: App updated.
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
...
  /deployedServiceFunctions:
    get:
      tags:
      - Service Functions Instances
      summary: Returns the edge service functions status per node.
      operationId: deployed_service_functions_status
      responses:
        "200":
          description: Returns the edge service functions status per node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deployedappsResponse'
        "405":
          description: Method not allowed
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
  /deployedServiceFunctions/{deployedServiceFunctionName}:
    get:
      tags:
      - Service Functions Instances
      summary: Returns the requested edge service status per node.
      operationId: deployed_service_function_status
      parameters:
      - name: deployedServiceFunctionName
        in: path
        description: Represents a  service function  from the running deployments
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Returns the requested edge service function status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deployedappsResponse'
        "405":
          description: Method not allowed
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
...
  /deployedServiceFunction/{deployedServiceFunctionName}:
    delete:
      tags:
      - Service Functions Instances
      summary: Deletes a deployed Service function.
      operationId: delete_deployed_service_function
      parameters:
      - name: deployedServiceFunctionName
        in: path
        description: Represents a service function from the running deployments.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Deployed service function  deleted
        "405":
          description: Method not allowed
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
...
  /chain/{chainServiceName}:
    delete:
      tags:
      - Service Functions Instances
      summary: Deletes a deployed chain.
      operationId: delete_chain
      parameters:
      - name: chainServiceName
        in: path
        description: Represents a chain Service  from the running deployments.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Deployed chain deleted
        "405":
          description: Method not allowed
      x-openapi-router-controller: swagger_server.controllers.service_functions_instances_controller
...
components:
  schemas:
...
    env_parameters:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        value_ref:
          type: string
      example:
        name: ELASTICSEARCH_URL
        value: http://elasticsearch:9200
...
    volume_mount_deploy:
      type: object
      properties:
          name:
            type: string
            example: volume1
          storage:
            type: string
            example: 100Mi
    serviceFunctionIndex:
      type: object
      properties:
          service_function_name:
            type: string
            example: mongodb
          service_function_index:
            type: integer
            example: 2
    deployServiceFunction:
      type: object
      properties:
        service_function_name:
          type: string
          example: mongodb
        service_function_instance_name:
          type: string
          example: mondodb_for_IoT_App
        volume_mounts:
          type: array
          items:
            $ref: '#/components/schemas/volume_mount_deploy'
        autoscaling_metric:
          type: string
          example: cpu
        autoscaling_policy:
          type: string
          example: minimize_cost
        count_min:
          type: integer
          example: 1
        count_max:
          type: integer
          example: 3
        location:
          type: string
          example: Peania_Athens_node1
        all_node_ports:
          type: boolean
          example: false
        node_ports:
          type: array
          items:
            type: integer
          example:
          - 90
          - 8080
        env_parameters:
          type: array
          items:
            $ref: '#/components/schemas/env_parameters'
...
    deployChain:
      type: object
      properties:
        chain_service_function_instance_name:
          type: string
        chain_service_function_order:
          type: array
          items:
            $ref: '#/components/schemas/serviceFunctionIndex'
        service_functions:
          type: array
          items:
            $ref: '#/components/schemas/deployServiceFunction'
...
    deployedappsResponse:
      type: object
      properties:
        nodeid:
          type: string
        nodename:
          type: string
        paasid:
          type: string
        paasname:
          type: string
        status:
          type: string
        exposedports:
          type: array
          items:
            type: integer
      example:
        paasname: paasname
        nodename: nodename
        exposedports:
        - 0
        - 0
        nodeid: nodeid
        paasid: paasid
        status: status
...
  securitySchemes:
    registry_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://serviceregistry.swagger.io/oauth/dialog
          scopes:
            write:services: modify services in your account
            read:services: read your services
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_registry_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_registry_auth
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key